Adicionei a função getTrips no TripsList.js
