+++ modulo3/projeto-labex/src/Components/TripsCard/TripsCard.js
@@ -2,11 +2,12 @@ import React from 'react'
 import {TripsCardContainer} from './style'
 
 
-export const TripsCard = () => {
+export const TripsCard = (props) => {
+
 
     return (
         <TripsCardContainer>
-            <p><strong>Nome:</strong></p>
+            <p><strong>Nome: </strong></p>
             <p><strong>Descrição:</strong></p>
             <p><strong>Planeta:</strong></p>
             <p><strong>Duração:</strong></p>
