+++ modulo3/projeto-labex/src/hooks/useRequestData.js
@@ -1,0 +1,15 @@
+// import {useEffect, useState} from 'react'
+// import axios from 'axios'
+
+// export function useRequestData(url) {
+
+//     const [data, setData] = useState(undefined)
+
+//     useEffect(() => {
+//         axios.get(url)
+//         .then((res) => setData(res.data))
+//         .catch((err) => console.log(err.response.data))
+//     }, [url])
+
+//     return data
+// }
